requirements.txt
tensorflow
numpy
pillow
pandas